renames test examples module and packages (prepare multi-example)
