+++ .idea/modules.xml
@@ -13,7 +13,7 @@
       <module fileurl="file://$PROJECT_DIR$/.idea/se.jbee.inject.convert.iml" filepath="$PROJECT_DIR$/.idea/se.jbee.inject.convert.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/se.jbee.inject.event.iml" filepath="$PROJECT_DIR$/.idea/se.jbee.inject.event.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/se.jbee.inject.lang.iml" filepath="$PROJECT_DIR$/.idea/se.jbee.inject.lang.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/test.example.app.iml" filepath="$PROJECT_DIR$/.idea/test.example.app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/test.examples.iml" filepath="$PROJECT_DIR$/.idea/test.examples.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/test.integration.iml" filepath="$PROJECT_DIR$/.idea/test.integration.iml" />
     </modules>
   </component>
