+++ .idea/test.integration.iml
@@ -14,7 +14,7 @@
     <orderEntry type="module" module-name="se.jbee.inject" scope="TEST" />
     <orderEntry type="module" module-name="se.jbee.inject.convert" scope="TEST" />
     <orderEntry type="module" module-name="se.jbee.inject.event" scope="TEST" />
-    <orderEntry type="module" module-name="test.example.app" scope="TEST" />
+    <orderEntry type="module" module-name="test.examples" scope="TEST" />
     <orderEntry type="module" module-name="se.jbee.inject.lang" scope="TEST" />
     <orderEntry type="module" module-name="se.jbee.inject.bootstrap" scope="TEST" />
     <orderEntry type="library" scope="TEST" name="bach-external-modules" level="project" />
