--- test.example.app/test/resources/META-INF/services/se.jbee.inject.bind.Bundle
@@ -1,2 +1,0 @@
-test.example.app.MyRootBundle
-test.example.app.MyEnvSetupBundle
