--- test.example.app/test/resources/META-INF/services/se.jbee.inject.bind.ModuleWith
@@ -1,1 +1,0 @@
-test.example.app.SupportAnnotationPattern
