+++ test.examples/test/java/test/example1/MyEnvSetupBundle.java
@@ -1,4 +1,4 @@
-package test.example.app;
+package test.example1;
 
 import se.jbee.inject.Env;
 import se.jbee.inject.Extends;
