+++ test.examples/test/resources/META-INF/services/se.jbee.inject.bind.Bundle
@@ -1,0 +1,2 @@
+test.example1.MyRootBundle
+test.example1.MyEnvSetupBundle
