+++ test.examples/test/resources/META-INF/services/se.jbee.inject.bind.ModuleWith
@@ -1,0 +1,1 @@
+test.example1.SupportAnnotationPattern
